Fix expression builder
