+++ canvas_modules/common-canvas/rollup.config.js
@@ -25,6 +25,9 @@ import url from "@rollup/plugin-url";
 import { visualizer } from "rollup-plugin-visualizer";
 
 const bundleReport = process.env.BUNDLE_REPORT;
+var window = {
+	navigator: {}
+};
 
 export default {
 	input: {
